service_sim/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
pydantic==2.9.2 
requests==2.32.3 
redis==6.0.0 
python-dotenv==1.0.1
PyYAML==6.0
confluent-kafka==2.10.0
psycopg2==2.9.10
faiss-cpu==1.11.0
markdown
sentence_transformers==4.1.0
numpy
asyncio